load theme on /doDaemonReady
